add vmvx-to-vm lit test and fix the pack.i8i8 import to use i32 type for padding_value

+++ compiler/src/iree/compiler/Dialect/VMVX/vmvx.imports.mlir
@@ -525,7 +525,7 @@ vm.import @pack.i8i8(
   %out_size1 : i64,
   %out_size2 : i64,
   %out_size3 : i64,
-  %padding_value : i8,
+  %padding_value : i32,
   %flags : i32
 )
